+++ src/main/kotlin/com/atlassian/performance/tools/virtualusers/EntryPoint.kt
@@ -1,11 +1,8 @@
 package com.atlassian.performance.tools.virtualusers
 
-import com.atlassian.performance.tools.jiraactions.ActionMetricStatistics
-import com.atlassian.performance.tools.jiraactions.MergingActionMetricsParser
 import com.atlassian.performance.tools.jiraactions.SeededRandom
 import com.atlassian.performance.tools.jiraactions.scenario.Scenario
 import com.atlassian.performance.tools.jirasoftwareactions.JiraSoftwareScenario
-import com.atlassian.performance.tools.report.PlaintextReport
 import com.atlassian.performance.tools.virtualusers.logs.LogConfigurationFactory
 import org.apache.commons.cli.*
 import org.apache.logging.log4j.LogManager
@@ -181,7 +178,7 @@ class Application {
             options.printHelp()
             System.exit(0)
         }
-        val metricsFiles = BasicTest(
+        BasicTest(
             jiraAddress = options.jiraAddress,
             scenario = options.scenario,
             adminLogin = options.adminLogin,
@@ -192,8 +189,5 @@ class Application {
             minimumRun = options.minimumRun,
             virtualUsers = options.virtualUsers
         )
-        val metrics = MergingActionMetricsParser().parse(metricsFiles)
-        val report = PlaintextReport(ActionMetricStatistics(metrics)).generate()
-        println(report)
     }
 }